push of proposed refactor
