+++ tests/integration/profiling/rule_based_profilers/test_rule_based_profiler.py
