--- tests/rule_based_profiler/alpha/README.md
@@ -1,1 +1,0 @@
-The test fixtures and modules, contained in this directory as well as in the directory "tests/test_fixtures/rule_based_profiler/alpha", are denoted as "alpha" to indicate that the Rule-Based Profiler (RBP), delivered on June 30, 2021, was implemented to specifications with high quality, but was not in the form that could be easily accessible by users.  The effort, commencing January 12, 2022 focuses on the RBP productization and fulfillment of the practical scenarios that are valueable to the user base.
