+++ tests/rule_based_profiler/test_rule.py
@@ -8,6 +8,7 @@ from great_expectations.rule_based_profiler.types import (
 )
 
 
+# <WILL> purely unit
 # noinspection PyPep8Naming
 def test_get_parameter_value_by_fully_qualified_parameter_name_invalid_parameter_name(
     column_Age_domain,
@@ -26,6 +27,7 @@ def test_get_parameter_value_by_fully_qualified_parameter_name_invalid_parameter
         )
 
 
+# <WILL> ok this test is insane
 # noinspection PyPep8Naming
 @pytest.mark.parametrize(
     "domain_name,fully_qualified_parameter_name,value,details,use_value_suffix,value_accessor,test_details,",
